+++ src/main/java/com/example/springapi/repositories/UserResponsitory.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserResponsitory extends JpaRepository<User, Long>{
+public interface UserResponsitory extends JpaRepository<User, Integer>{
     
 }
